Complete upgrade to Kinesis Client Library v2
