--- src/main/java/org/graylog/integrations/aws/transports/Proxy.java
@@ -1,18 +1,0 @@
-package org.graylog.integrations.aws.transports;
-
-import com.amazonaws.ClientConfiguration;
-import com.amazonaws.ClientConfigurationFactory;
-import okhttp3.HttpUrl;
-
-import javax.validation.constraints.NotNull;
-
-public class Proxy {
-
-    public static ClientConfiguration forAWS(@NotNull HttpUrl proxyUrl) {
-        return new ClientConfigurationFactory().getConfig()
-                                               .withProxyHost(proxyUrl.host())
-                                               .withProxyPort(proxyUrl.port())
-                                               .withProxyUsername(proxyUrl.username())
-                                               .withProxyPassword(proxyUrl.password());
-    }
-}
